+++ android/src/main/kotlin/com/fluttercandies/photo_manager/PhotoManagerPlugin.kt
@@ -86,6 +86,7 @@ class PhotoManagerPlugin : FlutterPlugin, ActivityAware {
             binding.addActivityResultListener(it.deleteManager)
             binding.addActivityResultListener(it.writeManager)
             binding.addActivityResultListener(it.favoriteManager)
+            binding.addActivityResultListener(it.pickerManager)
         }
     }
 
@@ -97,6 +98,7 @@ class PhotoManagerPlugin : FlutterPlugin, ActivityAware {
             oldBinding.removeActivityResultListener(p.deleteManager)
             oldBinding.removeActivityResultListener(p.writeManager)
             oldBinding.removeActivityResultListener(p.favoriteManager)
+            oldBinding.removeActivityResultListener(p.pickerManager)
         }
     }
 }